pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "amael-agentic-ia"
version = "1.0.0"
description = "Plataforma multi-agente modular para automatización inteligente"
requires-python = ">=3.11"

# ── Dependencias base (todas las capas de la plataforma) ──────────────────────
dependencies = [
    # Framework web
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-multipart>=0.0.9",

    # HTTP client
    "httpx>=0.27.0",
    "requests>=2.31.0",

    # LangChain + LangGraph (Ollama — proveedor único activo)
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-ollama>=0.1.0",
    "langchain-qdrant>=0.1.0",
    "langchain-text-splitters>=0.2.0",
    "langgraph>=0.1.0",

    # Autenticación
    "authlib>=1.3.0",
    "python-jose[cryptography]>=3.3.0",
    "itsdangerous>=2.1.0",

    # Almacenamiento
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.4",
    "minio>=7.2.7",
    "qdrant-client>=1.9.0",

    # Observabilidad
    "prometheus-client>=0.20.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-instrumentation-httpx>=0.45b0",
    "opentelemetry-instrumentation-requests>=0.45b0",

    # Utilidades
    "python-magic>=0.4.27",
    "Pillow>=10.3.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
# Agente SRE / k8s-agent
sre = [
    "kubernetes>=29.0.0",
    "apscheduler>=3.10.4",
]

# Procesamiento de documentos (ResearchAgent)
docs = [
    "pypdf>=4.2.0",
    "python-docx>=1.1.0",
    "duckduckgo-search>=6.1.0",
]

# TTS (CosyVoice)
tts = [
    "torch>=2.3.0",
]

# Desarrollo y tests
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",        # cliente para TestClient de FastAPI
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

# Todas las dependencias opcionales juntas
all = [
    "amael-agentic-ia[sre,docs,dev]",
]

[tool.setuptools.packages.find]
where = ["."]
include = [
    "core*", "config*", "observability*", "security*",
    "storage*", "agents*", "skills*", "tools*",
    "orchestration*", "llm*", "memory*", "interfaces*",
]

# ── Herramientas de desarrollo ────────────────────────────────────────────────

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]  # line too long — cubierto por line-length

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false
